fix(csv): display full media URLs in the CSV . Move special filesystem URL logic to service provider
